#758: Move HTTPClient live value to TDFoundation
