+++ ToDo-Garden/ToDo-Garden/AppCore/Sources/AppCore/StaticObjects.swift
@@ -11,29 +11,6 @@ import HTTPClient
 import TDFoundation
 import TimerScene
 
-extension HTTPClient {
-  static let live = HTTPClient(
-    transport: URLSessionTransport(urlSession: URLSession.shared),
-    middlewares: [
-      AuthenticationMiddleWare(),
-      CommonHeaderMiddleware(),
-      RefreshMiddleware(accessTokenManager: AccessTokenManager.live)
-    ]
-  )
-}
-
-extension AccessTokenManager {
-  static let live = AccessTokenManager(
-    httpClient: HTTPClient(
-      transport: URLSessionTransport(urlSession: URLSession.shared),
-      middlewares: [
-        AuthenticationMiddleWare(),
-        CommonHeaderMiddleware()
-      ]
-    )
-  )
-}
-
 extension TimerSceneSceneBuilder.Dependency {
   @MainActor
   public static let live = TimerSceneSceneBuilder.Dependency(
